Support for IP Addresses
Note that variables are replaced with ip addresses starting with 0.0.0.1. If there are three variables, "x","y" and "z", then "x" is replaced by 0.0.0.1, "y" is replaced by 0.0.0.2, and "z" is replaced by 0.0.0.3. We have to make sure that these IP Addresses do not collide with any IP address in the